add bug: disabling plugin after moving user list
If the user list is moved while plugin is enabled, then disabling
the plugin does not remove the channel tree. This bug existed before
the previous commit, which made it possible to move the user list
while plugin is enabled, but it was not discovered since moving the
user list itself was broken.