add menu items
Contributed by Robert Utasi (hunboy)

This is menu items for channel tree. I've put it into the init() function
because the channel tree plugin itself can be enabled/disabled via menu,
so that is permanently there. I've wrapped the client.updateMenus()
functions at the enable() and disable() functions into the setTimeout(),
because if an active menu is opened when the updateMenus() function runs
over of that, that menu goes to damaged. Currently the view menu can be
damaged without that wrap.