workaround for two bugs related to tree styling
Contributed by Robert Utasi (hunboy)

The bug#2 happens not only when that is manually loaded, that happens
all the time, just the situation is different a little. So the bug#1 and
the bug#2 is totally same, just one for the previously loaded items and
one for the lately loaded items. An this is interesenting, because the
situation is a little bit of different in gecko2.0 than gecko1.9.2. but
still buggy. there is only 1 case when permanently good, when you replace
the tree. I ve tested everything reordering the stuffs in the enable()
function but still buggy. So seems it simply hates when you insert that
tree to the DOM in the enable() function. But there is a question -
because the replace is good - what is that minimum delay when it should
good already. I found a few workarounds and this one looks the most less
ugly. The delay depends on the client's processor speed, and depends on
the other loaded plugins at startup (those needs time also). This 300ms
seems stable in the different versions and computers, and this is still
inside the human's reactiontime, so almost invisible. The workaround
is advisedly at the end of the enable() funtion because of the time
reasons. I think this is a view-rendering core bug, and we can't fix it
via plugin with a real fix, so this is a working workaround.