correct the display of top level leaf nodes
Contributed by Robert Utasi (hunboy)

This one is for hiding the container when that subtree is empty. The
bug#4. The modification is: this is not container at startup, only later
when an item arrives into that. During closing a view the container attr
will be removed when that was the last item which one closed.